Add files via upload
This is a Matlab file last compiled 24 November 2017. It is the first draft for the single dimensional case of an electron moving through a lattice of atoms. What this code does thus far is solve the time-independent Schroedinger equation by setting up the Hamiltonian (matrix) and solving for the eigenvalues (E's, or allowed energies).